requirements.txt

 torchvision
 matplotlib
 pytorch-lightning
 wandb
 timm
 easydict
 opencv-python
 scipy
 scikit-learn 